+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 - Feature: Standard calendar, Horizontal  calendar
 
 ## 1.0.1
-- New Feature: Support the multiple selected  the same month.
-- New Feature: Support set default start selected and end selected the same month.
+- New Feature: Support the multiple select  the same month.
+- New Feature: Support range select with default start select and end select the same month.
+- New Feature: Support set default single select date.